Change for Linux to builtin printf
This suggested PR changes the order of which printf is being used in
`choose_printf()` so that Linux uses not /usr/bin/printf .
This is for some checks slightly faster.

Also PRINTF is now a tunable parameter which can be set through the enviroment
like PRINTF=/usr/bin/printf ./testssl.sh <cmdline>

It doesn't / shouldn't solve the bug #2783. But it might help to reproduce it.