Fix missing issuer CN
This fixes a problem which was introduced @ 8d8f83ace507db6a699acb4901d1329f31731a04.
It caused for some hosts not to parse / display the issuer CN correctly.

Also it adds some code in testssl.sh and in a unit test to detect
this earlier. In general an output string FIXME will now cause a
unit test to fail. This can + should be used at other places too!

Fixes #2789