Opossum vulnerabilty
Fixes #2833

This does a check for the opossum vulnerability, see https://opossum-attack.com/.

Currently it uses wget or curl, so one if them has to be present.

Proxy handling was introduced in check_pwnedkeys() which should help
that function too.

Also timeouts were used for curl and wget for better maturity.

Todos:
- handling when neither curl or wget is present
- at least a warning when no HTTP service is present
- take care of the diffs in http_get_header() to make this
  work --> make sure it still does using check_pwnedkeys()
  - -H ''?
  - ret codes
- backport proxy stuff in http_get_header() to 3.2
- backport curl/wget timeouts to 3.2
- manpage (incl. RFC 2817)